refactor: Showing 객체의 위치 view로 이동
