+++ README.md
@@ -51,7 +51,7 @@
 - [x] Movie | UiModel 만들기
 - [x] MovieItem | Movie 네이밍 변경
 - [x] ScheduleTime | 메서드 분리
-- [ ] Showings | 스스로 일을 할 수 있게 변경
+- [x] Showings | 스스로 일을 할 수 있게 변경
 - [ ] Ticket | 15 값 enum class로 변경
 - [x] MovieSchedule | filter를 toList 전에 호출되도록 변경
 - [x] MainActivity | commit 하는 부분 공통부분 메서드 분리
@@ -64,5 +64,5 @@
 - [ ] HomeFragment | superCall의 필요성?
 - [x] MovieAdapter |  areItemsTheSame | 아이템을 비교하는 메서드 수정
 - [x] TheaterBottomSheetDialogFragment | 생성자 삭제
-- [ ] 데이터 바인딩
+- [x] 데이터 바인딩
 - [x] MainActivity | 화면이 회전되어도 현재 Fragment가 유지되도록 수정
