+++ app/src/main/java/woowacourse/movie/view/home/theater/OnTheaterEventListener.kt
@@ -1,7 +1,5 @@
 package woowacourse.movie.view.home.theater
 
-import woowacourse.movie.domain.Showing
-
 fun interface OnTheaterEventListener {
     fun onClickReservation(showing: Showing)
 }
