+++ app/src/main/res/layout/fragment_theater_bottom_sheet_dialog.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="showings"
-            type="java.util.List&lt;woowacourse.movie.domain.Showing>" />
+            type="java.util.List&lt;woowacourse.movie.view.home.theater.Showing>" />
 
         <variable
             name="onItemClick"
