+++ app/src/main/res/layout/theater_item.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="showing"
-            type="woowacourse.movie.domain.Showing" />
+            type="woowacourse.movie.view.home.theater.Showing" />
 
         <variable
             name="onConfirm"
